+++ analysis/tests/resources/dummy_tool.sh
@@ -7,3 +7,5 @@ echo "SUT: $2"
 echo "test file" > file1
 mkdir -p dir1
 echo "test file" > dir1/file2
+echo "EnvVar1: $EnvVar1"
+echo "EnvVar2: $EnvVar2"